tests on the backend side are merged
