+++ backend/partial_update.log
@@ -1,0 +1,3 @@
+first_name: John
+last_name: Doe
+email: tester@gmail.com
